refactor: lint files refactor: move extractQueryFromContext to separate file
