+++ src/components/ItemsPerPageSelect.tsx
@@ -1,4 +1,3 @@
-
 import { useRouter } from 'next/router';
 import { useAppSelector } from '../redux/hooks';
