+++ src/pages/404.tsx
@@ -1,4 +1,4 @@
-import Link from "next/link";
+import Link from 'next/link';
 
 export default function NotFoundPage() {
   return (
