+++ src/pages/_app.tsx
@@ -1,9 +1,9 @@
-import { wrapper } from '@/redux/store'
-import type { AppProps } from 'next/app'
+import { wrapper } from '@/redux/store';
+import type { AppProps } from 'next/app';
 import { FC } from 'react';
 import { Provider } from 'react-redux';
 
-import '@/styles/globals.css'
+import '@/styles/globals.css';
 
 const App: FC<AppProps> = ({ Component, ...rest }) => {
   const { store, props } = wrapper.useWrappedStore(rest);
